1. Fix swscale initiation failures when MMX is used. summary of issue: upcasting negative numbers to int64_t and giving these to swscale = angry swscale. 2. swscale has started some SSE2 codepaths, add SSE2 support.
Originally committed to SVN as r405.

+++ src/avisynth/avsutils.h
@@ -28,7 +28,7 @@ extern "C" {
 #include "ffmscompat.h"
 
 
-int AvisynthToFFCPUFlags(long AvisynthFlags);
+int64_t AvisynthToFFCPUFlags(long AvisynthFlags);
 PixelFormat CSNameToPIXFMT(const char *CSName, PixelFormat Default);
 int ResizerNameToSWSResizer(const char *ResizerName);
